+++ services/unique_id/warehouse.go
@@ -8,7 +8,7 @@ type WareHouse struct {
   HouseMap map[string]*IdSet // every business type has an IdSet
 }
 
-func (w *WareHouse) Acquire(business string) (int, error){
+func (w *WareHouse) Acquire(business string) (uint64, error){
   businessSet, ok := w.HouseMap[business]
   if ok == false{
     return 0, errors.New("no business type")
